Add remaining trait files and update configs for comprehensive improvements
