+++ README.md
@@ -33,46 +33,45 @@ $user->setPhone(
 self::assertSame(expected: $phoneAfter, actual: $user->getPhone());
 self::assertNotSame(expected: $phoneBefore, actual: $user->getPhone());
 
-// fluent assertions
-fact($user->getPhone())
-    ->is($phoneAfter)
-    ->equals($phoneAfter)
-    ->not($phoneBefore)
-    ->true()
-    ->notTrue()
-    ->false()
-    ->notFalse()
-    ->null()
-    ->notNull()
-    ->matchesRegularExpression('#^\d+$#')
-    ->notMatchesRegularExpression('#^\D+$#')
-    ->containsString('alpha')
-    ->notContainsString('alpha')
-    ->containsStringIgnoringCase('beta')
-    ->notContainsStringIgnoringCase('beta')
-    ->count(5)
-    ->notCount(5)
-    ->arrayHasKey('echo')
-    ->arrayNotHasKey('echo')
-    ->instanceOf(UserFactory::class)
-    ->notInstanceOf(UserFactory::class)
-    ->ulid() // Universally Unique Lexicographically Sortable Identifier https://github.com/ulid/spec
-    ->isLowerThan(100)
-    ->isGreaterThan(50)
-    ->isPositive()
-    ->isNegative()
-    ->isZero()
-    ->isBetween(1, 10)
-    ->startsWith('prefix')
-    ->endsWith('suffix')
-    ->hasLength(5)
-    ->contains('value')
-    ->doesNotContain('value')
-    ->hasSize(3)
-    ->isEmptyArray()
-    ->isEmptyString()
-    ->isInt()
-    ->isString()
+### Comparison and Equality Methods
+```php
+fact(42)->is(42)->equals(42)->not(43);
+```
+
+### Boolean Methods
+```php
+fact(true)->true()->notFalse();
+```
+
+### Null Methods
+```php
+fact(null)->null()->notNull();
+```
+
+### Numeric Methods
+```php
+fact(5)->isLowerThan(10)->isGreaterThan(0)->isPositive()->isNegative()->isZero()->isBetween(1, 10);
+```
+
+### String Methods
+```php
+fact('hello world')->matchesRegularExpression('/\w+/')->startsWith('hello')->endsWith('world')->hasLength(11)->isEmptyString();
+```
+
+### Array Methods
+```php
+fact([1, 2, 3])->count(3)->contains(2)->doesNotContain(4)->hasSize(3)->isEmptyArray();
+```
+
+### Type Checking Methods
+```php
+fact(42)->isInt()->isString()->instanceOf(int::class)->hasProperty('name')->hasMethod('doSomething');
+```
+
+### Special Methods
+```php
+fact('01ARZ3NDEKTSV4RRFFQ69G5FAV')->ulid();
+```
     ...
     ;
