+++ src/Traits/SpecialAssertions.php
@@ -1,0 +1,28 @@
+<?php
+
+declare(strict_types=1);
+
+namespace K2gl\PHPUnitFluentAssertions\Traits;
+
+use K2gl\PHPUnitFluentAssertions\Reference\RegularExpressionPattern;
+
+trait SpecialAssertions
+{
+    /**
+     * Asserts that a variable is a valid ULID.
+     *
+     * This method checks if the actual value matches the ULID (Universally Unique Lexicographically Sortable Identifier) format.
+     *
+     * @see https://github.com/ulid/spec
+     *
+     * Example usage:
+     * fact('01ARZ3NDEKTSV4RRFFQ69G5FAV')->ulid(); // Passes (if valid ULID)
+     * fact('invalid-ulid')->ulid(); // Fails
+     *
+     * @return self Returns the FluentAssertions instance for method chaining.
+     */
+    public function ulid(): self
+    {
+        return $this->matchesRegularExpression(RegularExpressionPattern::ULID);
+    }
+}
